requirements.txt
# 基础依赖
numpy>=1.26.0
pydantic>=2.7.4
fastapi>=0.110.0
uvicorn>=0.27.1
python-multipart>=0.0.9
requests>=2.31.0
tqdm>=4.66.0
httpx[socks]>=0.27.0
psutil>=5.9.0  # 用于系统资源监控

# LangChain 相关
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-huggingface>=0.0.10

# 向量数据库
qdrant-client>=1.7.0

# 模型相关
transformers>=4.38.2
torch>=2.2.0
sentence-transformers>=2.5.1
accelerate>=0.27.0
bitsandbytes>=0.42.0
huggingface-hub>=0.20.0

# 本地模型支持
llama-cpp-python>=0.2.0

# 文档处理
python-docx>=1.1.0
python-pptx>=0.6.23
PyPDF2>=3.0.0

# Web 界面
streamlit>=1.32.0

# 开发工具
black>=24.2.0  # 代码格式化
isort>=5.13.0  # import 排序
mypy>=1.8.0   # 类型检查
pytest>=8.0.0  # 单元测试